securefiles/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pyotp
gunicorn
